Added summary to evaluate_tflite.py and moved method to evaluate_tools.py.
