+++ evaluate_tflite.py
@@ -15,8 +15,8 @@ from six.moves import zip, range
 from multiprocessing import JoinableQueue, Process, cpu_count, Manager
 from deepspeech import Model
 
-from util.evaluate_tools import calculate_report
-from util.flags import create_flags
+from util.evaluate_tools import calculate_report, print_report
+from util.flags import create_flags, FLAGS
 
 r'''
 This module should be self-contained:
@@ -98,11 +98,14 @@ def main(args, _):
         predictions.append(msg['prediction'])
         wavlist.append(msg['wav'])
 
+    # Print test summary
     wer, cer, samples = calculate_report(wav_filenames, ground_truths, predictions, losses)
     mean_loss = np.mean(losses)
+    print('Test - WER: %f, CER: %f, loss: %f' % (wer, cer, mean_loss))
+    print('-' * 80)
 
-    print('Test - WER: %f, CER: %f, loss: %f' %
-          (wer, cer, mean_loss))
+    # Print some examples
+    print_report(samples)
 
     if args.dump:
         with open(args.dump + '.txt', 'w') as ftxt, open(args.dump + '.out', 'w') as fout:
